Calculates first-order statistical metrics for RIA_image

Calculates first-order statistical metrics of RIA_image. First-order metrics discard all spatial information. By default the $modif image will be used to calculate statistics. If use_slot is given, then the data present in RIA_image$use_slot will be used for calculations. Results will be saved into the $stat_fo slot. The name of the subslot is determined by the supplied string in $save_name, or is automatically generated by RIA.

first_order(RIA_data_in, use_type = "single", use_orig = TRUE,
  use_slot = NULL, save_name = NULL, verbose_in = TRUE)

Arguments

RIA_data_in
RIA_image.

use_type
string, can be "single" which runs the function on a single image, which is determined using "use_orig" or "use_slot". "discretized" takes all datasets in the RIA_image$discretized slot and runs the analysis on them.

use_orig
logical, indicating whether to use image present in RIA_data$orig. If FALSE, the modified image will be used stored in RIA_data$modif.

use_slot
string, name of slot where data wished to be used is. Use if the desired image is not in the data$orig or data$modif slot of the RIA_image. For example, if the desired dataset is in RIA_image$discretized$ep_4, then use_slot should be discretized$ep_4. The results are automatically saved. If the results are not saved to the desired slot, then please use save_name parameter.

save_name
string, indicating the name of subslot of $stat_fo to save results to. If left empty, then it will be automatically determined.

verbose_in
logical indicating whether to print detailed information. Most prints can also be suppresed using the suppressMessages function.

Value

RIA_image containing the statistical information.

References

Márton KOLOSSVÁRY et al. Radiomic Features Are Superior to Conventional Quantitative Computed Tomographic Metrics to Identify Coronary Plaques With Napkin-Ring Sign Circulation: Cardiovascular Imaging (2017). DOI: 10.1161/circimaging.117.006843 https://www.ncbi.nlm.nih.gov/pubmed/29233836

Márton KOLOSSVÁRY et al. Cardiac Computed Tomography Radiomics: A Comprehensive Review on Radiomic Techniques. Journal of Thoracic Imaging (2018). DOI: 10.1097/RTI.0000000000000268 https://www.ncbi.nlm.nih.gov/pubmed/28346329

Examples